feat(gigya/loader): remove domain connector
