+++ components/gigya/loader/package.json
@@ -10,7 +10,6 @@
   },
   "dependencies": {
     "@s-ui/component-dependencies": "1",
-    "@s-ui/react-domain-connector": "1",
     "@schibstedspain/sui-script-loader": "1"
   },
   "keywords": [],
